requirement.txt
numpy
scipy
matplotlib
ipython
jupyter
pandas
sympy
nose
vtk
pyoctree
numpy-stl